Move the description text flow container info more general location Because it will be used in other editor.
