+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Components/Markdown/IDescriptionAction.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Localisation;
+
+namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Markdown;
+
+public interface IDescriptionAction
+{
+    public LocalisableString Text { get; set; }
+}
